social_distribution/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
Django==4.1.2
dj-database-url==1.0.0
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-on-heroku==1.1.2
djangorestframework==3.14.0
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
psycopg2-binary==2.9.4
pycodestyle==2.9.1
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.5
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
soupsieve==2.3.2.post1
sqlparse==0.4.3
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
cmarkgfm